Move file comparators to separated public classes
